% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curse.R
\name{curse}
\alias{curse}
\title{Curse combinations}
\usage{
curse(x = 3)
}
\arguments{
\item{x}{integer specifying number of curse words to return, default is 3}
}
\description{
replicates 'curse_grab' helper script specifying number of curse words to output
}
