% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curser-package.R
\docType{package}
\name{curser}
\alias{curser}
\alias{curser-package}
\title{curser package}
\description{
Express anger or frustration in code
}
\author{
\strong{Maintainer}: Alessandro Gentile \email{alessandro.gentile03@gmail.com}

}
